Apply some codestyle notes and rename the plugin
