+++ docker/plugins/bmqprometheus/Dockerfile
