+++ docker/plugins/bmqprometheus/docker-compose.yml
