+++ src/plugins/bmqprometheus/build_prometheus_deps.sh
