+++ src/plugins/bmqprometheus/package/bmqprometheus.mem
@@ -1,0 +1,5 @@
+bmqprometheus_entry
+bmqprometheus_pluginlibrary
+bmqprometheus_prometheusstatconsumer
+bmqprometheus_version
+bmqprometheus_versiontag
