+++ src/plugins/bmqprometheus/tests/docker/Dockerfile
