+++ src/plugins/bmqprometheus/tests/docker/docker-compose.yml
