+++ src/plugins/bmqprometheus/tests/docker/prometheus.yml
