+++ src/plugins/bmqprometheus/tests/prometheus_localhost.yaml
