--- src/plugins/prometheus/package/prometheus.mem
@@ -1,5 +1,0 @@
-prometheus_entry
-prometheus_pluginlibrary
-prometheus_prometheusstatconsumer
-prometheus_version
-prometheus_versiontag
